Moved Files
~After a lot of errors I realized that we were putting all of our
homemade files into the internal folder. This was right for last year’s
code but wrong for this year’s code
~due to a lot of changes made by the normal first files and the game
manual, we were forced to delete some files and import the remaining
ones into the team code folder. I hope this is the correct thing to do
and it won’t mess up any of the grade builds or file paths